% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{ts_accuracy}
\alias{ts_accuracy}
\title{Time series accuracy}
\usage{
ts_accuracy(ts_df, lag = 1)
}
\arguments{
\item{ts_df}{A tibble containing dates in column labeled 'ds' and values in column labeled 'y'.
The point forecast should be in a column labeled 'yhat', and respective prediction intervals should
be in columns labeled 'yhat_lower' and 'yhat_upper'.}

\item{lag}{The lag to be used in calculating MASE, which should be the period of seasonality of the time series.}
}
\value{
A tibble with ME, MAE, MASE, Coverage, and Width columns.
}
\description{
Calculates ME, MAE, MASE, Coverage, and Width for a time series tibble.
}
\examples{
df <- tidyr::tibble(
  ds = lubridate::today() - 100:1,
  y = rnorm(100),
  yhat = rnorm(100),
  yhat_lower = yhat - 1,
  yhat_upper = yhat + 1
)
ts_accuracy(df)
}
