% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_graph.R
\name{ts_graph}
\alias{ts_graph}
\title{Graph time series}
\usage{
ts_graph(ts_df)
}
\arguments{
\item{ts_df}{A tibble containing dates in column labeled 'ds' and
values in column labeled 'y'.
If forecast is contained, the point forecast should be in a column labeled 'yhat', and
respective prediction intervals should be in columns labeled 'yhat_lower' and 'yhat_upper'.}
}
\value{
An interactive graph.
}
\description{
Create an interactive graph of a time series tibble.
}
\examples{
ts_graph(tidyr::tibble(ds = lubridate::today() - 100:1, y = rnorm(100)))
}
