% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seasonality.R
\name{ts_seasonality}
\alias{ts_seasonality}
\title{Explore time series seasonality}
\usage{
ts_seasonality(ts_df, ts_type)
}
\arguments{
\item{ts_df}{A tibble containing dates in column labeled 'ds' and values in column labeled 'y'.}

\item{ts_type}{A string indicating the type of time series. Valid options are 'daily', 'monthly', and 'quarterly'.}
}
\value{
A seasonality plot.
}
\description{
Creates side-by-side boxplots for analyzing seasonality for weekly, monthly, and quarterly time series tibbles.
}
\examples{
df <- tidyr::tibble(ds = lubridate::today() - 100:1, y = rnorm(100))
ts_seasonality(df, 'daily')
}
