% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seasonality.R
\name{ts_seasonality}
\alias{ts_seasonality}
\title{Explore time series seasonality}
\usage{
ts_seasonality(ts_df, seasonalities = c("weekly", "monthly", "yearly"))
}
\arguments{
\item{ts_df}{A tibble containing dates in column labeled 'ds' and values in column labeled 'y'.}

\item{seasonalities}{A string containing which seasonalities will be plotted. The options are 'weekly',
'monthly', and 'yearly'.}
}
\value{
A collection of seasonality plots.
}
\description{
Creates side-by-side boxplots for analyzing weekly, monthly, and yearly seasonalities for
a time series tibble.
}
\examples{
df <- tidyr::tibble(ds = lubridate::today() - 100:1, y = rnorm(100))
ts_seasonality(df)
}
